Refactored class definitions in addDialogueWindow.py module to seperate modules
